+++ src/lib/EventOverlay/index.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import ReactDOM from 'react-dom';
+import { omit } from 'lodash';
 
 export default class EventOverlay extends React.Component {
   static displayName = 'EventOverlay';
@@ -23,7 +24,14 @@ export default class EventOverlay extends React.Component {
       ||
       prevProps.direction !== this.props.direction
     ) {
+      this.addKeyHandlers();
       return this.forceUpdate(() => this.isVisible());
+    } else if (
+      !this.props.isOpen
+      &&
+      prevProps.isOpen !== this.props.isOpen
+    ) { 
+      return this.removeKeyHandlers();
     }
   };
 
@@ -146,7 +154,6 @@ export default class EventOverlay extends React.Component {
 
     if(allowClickAway) {
       document.addEventListener('click', this.handleAllowClickAway, true);
-      document.addEventListener('keyup', this.handleKeyUp, true);
     }
 
     closeOnClick && document.addEventListener('click', this.handleCloseOnClick, false);
@@ -156,13 +163,24 @@ export default class EventOverlay extends React.Component {
     this.isVisible();
   };
 
+  addKeyHandlers = () => {
+    const { allowClickAway } = this.props;
+
+    if(allowClickAway) {
+      document.addEventListener('keydown', this.handleKeyDown, false);
+    }
+  }
+
+  removeKeyHandlers = () => {
+    document.removeEventListener('keydown', this.handleKeyDown, false);
+  }
+
   removeHandlers = () => {
     document.removeEventListener('click', this.handleAllowClickAway, true);
     document.removeEventListener('click', this.handleCloseOnClick, false);
 
     window.removeEventListener('resize', this.handleResize, true);
     document.removeEventListener('scroll', this.handleScroll, false);
-    document.removeEventListener('keyup', this.handleKeyUp, true);
   };
 
   handleCloseOnClick = e => {
@@ -176,9 +194,10 @@ export default class EventOverlay extends React.Component {
     );
   };
 
-  handleKeyUp = e => {
+  handleKeyDown = e => {
     if (!this.props.isOpen) return;
     if (e.keyCode === 27) return this.handleClickAway(e);
+
     const anchorNode = ReactDOM.findDOMNode(this.props.anchorNode);
 
     return (
@@ -355,8 +374,29 @@ export default class EventOverlay extends React.Component {
   };
 
   render() {
-    const { className, isOpen, children, showArrow, maxHeight, maxWidth, style } = this.props;
+    const {
+      children,
+      className,
+      isOpen,
+      maxHeight,
+      maxWidth,
+      showArrow,
+      style,
+      ...props
+    } = this.props;
     const side = this.state.visibleDirection.split('-')[0];
+
+    const otherProps = omit({...props}, [
+      'allowClickAway',
+      'anchorNode',
+      'checkOverflow',
+      'close',
+      'closeOnClick',
+      'direction',
+      'isDynamic',
+      'targetOffset',
+    ]);
+
     const contentNodes = (
       <div
          className={
@@ -380,6 +420,7 @@ export default class EventOverlay extends React.Component {
             ...maxHeight && { maxHeight: `${maxHeight}px` },
             ...style
           }}
+          {...otherProps}
         >
           {children}
         </div>