+++ src/lib/ListItemSection/index.js
@@ -7,8 +7,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-const ListItemSection = props => { 
-    const { children, position, className } = props;
+const ListItemSection = props => {
+    const {
+      children,
+      className,
+      position,
+    } = props;
 
   return (
     <div
@@ -23,15 +27,15 @@ const ListItemSection = props => {
 };
 
 ListItemSection.propTypes = {
-  position: PropTypes.oneOf(['left', 'center', 'right', 'center-align']),
   children: PropTypes.node,
-  className: PropTypes.string
+  className: PropTypes.string,
+  position: PropTypes.oneOf(['left', 'center', 'right', 'center-align']),
 };
 
 ListItemSection.defaultProps = {
-  position: 'center',
   children: null,
-  className: ''
+  className: '',
+  position: 'center',
 };
 
 ListItemSection.displayName = 'ListItemSection';